cloud-run-ai-service-backed/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
google-generativeai>=0.8.0
google-auth==2.28.1
